Dispatcher: Datenbank Konfiguration

+++ FortDay-Dispatcher-ejb/ejbModule/de/fh_dortmund/inf/cw/surstwalat/dispatcher/beans/repositories/ActionResultRepositoryBean.java
@@ -1,14 +1,15 @@
 package de.fh_dortmund.inf.cw.surstwalat.dispatcher.beans.repositories;
 
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.domain.ActionResult;
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces.ActionResultRepositoryLocal;
 import javax.ejb.Stateless;
+
+import de.fh_dortmund.inf.cw.surstwalat.common.model.ActionResult;
+import de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces.ActionResultRepositoryLocal;
 
 /**
  * Session Bean implementation class ActionResultRepositoryBean
  */
 @Stateless
-public class ActionResultRepositoryBean extends DispatcherRepository<ActionResult, Long> implements ActionResultRepositoryLocal {
+public class ActionResultRepositoryBean extends DispatcherRepository<ActionResult, Integer> implements ActionResultRepositoryLocal {
 
     /**
      * Default constructor.
