+++ FortDay-Dispatcher-ejb/ejbModule/de/fh_dortmund/inf/cw/surstwalat/dispatcher/interfaces/ActionRepositoryLocal.java
@@ -2,9 +2,9 @@ package de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces;
 
 import javax.ejb.Local;
 
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.domain.Action;
+import de.fh_dortmund.inf.cw.surstwalat.common.model.Action;
 
 @Local
-public interface ActionRepositoryLocal extends IDispatcherRepository<Action, Long> {
+public interface ActionRepositoryLocal extends IDispatcherRepository<Action, Integer> {
 
 }
