+++ FortDay-Dispatcher-ejb/ejbModule/de/fh_dortmund/inf/cw/surstwalat/dispatcher/interfaces/ActionResultRepositoryLocal.java
@@ -2,9 +2,9 @@ package de.fh_dortmund.inf.cw.surstwalat.dispatcher.interfaces;
 
 import javax.ejb.Local;
 
-import de.fh_dortmund.inf.cw.surstwalat.dispatcher.domain.ActionResult;
+import de.fh_dortmund.inf.cw.surstwalat.common.model.ActionResult;
 
 @Local
-public interface ActionResultRepositoryLocal extends IDispatcherRepository<ActionResult, Long> {
+public interface ActionResultRepositoryLocal extends IDispatcherRepository<ActionResult, Integer> {
 
 }
